TST: and the tests for fit2d are good, although the real test is making a mask and trying to load it
